webcam-example/requirements.txt
absl-py==0.2.0
appnope==0.1.0
astor==0.6.2
backcall==0.1.0
bleach==1.5.0
click==6.7
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
Flask==0.12.2
Flask-SocketIO==2.9.4
gast==0.2.0
greenlet==0.4.13
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.1.6
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.1
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
pstats-view==0.1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-engineio==2.0.2
python-socketio==1.8.4
pytz==2018.4
PyYAML==3.12
pyzmq==17.0.0
qgrid==1.0.2
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.1
